aqi_weather_location_tests/requirements.txt
appdirs==1.4.4
arrow==0.17.0
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
cssselect==1.1.0
cssutils==1.0.2
distlib==0.3.1
filelock==3.0.12
idna==2.10
importlib-metadata==3.4.0
lxml==4.6.2
premailer==3.7.0
pyaml==20.4.0
pyjson5==1.4.8
PyMySQL==1.0.2
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.6.0
six==1.15.0
times==0.7
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.4.0
yagmail==0.14.245
zipp==3.4.0
